Move SocialMedia into the entity package
