+++ pom.xml
@@ -16,7 +16,6 @@
           </excludes>
         </configuration>
         <groupId>org.springframework.boot</groupId>
-        <version>2.5.4</version>
       </plugin>
       <plugin>
         <artifactId>spring-boot-maven-plugin</artifactId>
@@ -122,7 +121,7 @@
     </dependency>
     <dependency>
       <groupId>io.springfox</groupId>
-      <artifactId>springfox-swagger2 </artifactId>
+      <artifactId>springfox-swagger2</artifactId>
       <version>3.0.0</version>
     </dependency>
     <dependency>
